Improve release.ls output
The output of release.ls is now formatted as a human-readable table, and
it now accepts the --contains parameter (similar to projects.ls) that
adds an additional field to the output stating whether that release
contains the specified commit.

This also neatens fatal logging, and adds support for git revparse in
places that previously only accepted a full 40 char commit hash.